requirements.txt
requests
dotenv
webbrowser
flask
PyOpenSSL
